requirements.txt
requests
python-dotenv
fastapi
uvicorn
python-multipart
supabase
